split Dao interface into smaller interfaces and moved to new package
