+++ app/src/main/java/ua/blackwind/limbushelper/data/db/dao/Dao.kt
@@ -1,0 +1,6 @@
+package ua.blackwind.limbushelper.data.db.dao
+
+import androidx.room.Dao
+
+@Dao
+interface Dao: SinnerDao, IdentityDao, SkillDao, PartyDao
